fix pr comment and small refactoring
