+++ e2e/src/e2e/auth.setup.spec.ts
@@ -1,8 +1,8 @@
 import { test as setup  } from "@playwright/test";
-import { startPageDataTestId, loginPageDataTestId } from "../support/e2e";
+import { startPageDataTestId, loginPageDataTestId } from "../../support/e2e";
 import 'dotenv/config';
 import { get } from "http";
-import { USER_AUTH_STORAGE_PATH, userName, password } from "../../playwright.config";
+import { USER_AUTH_STORAGE_PATH, userName, password } from "../../../playwright.config";
 
 setup('login to site', async ({page, context}) =>{
     await page.goto('/');
